srv/deploy/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
httpx==0.28.1
websockets==14.1
docker==7.1.0
